// frontend/src/api/models/deltaSyncRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { MemberChangesDto } from "./memberChangesDto";
import type { ExpenseChangesDto } from "./expenseChangesDto";
import type { ExpenseItemChangesDto } from "./expenseItemChangesDto";
import type { SettlementChangesDto } from "./settlementChangesDto";
import type { BillMetaChangesDto } from "./billMetaChangesDto";

export interface DeltaSyncRequest {
  baseVersion?: number;
  members?: MemberChangesDto;
  expenses?: ExpenseChangesDto;
  expenseItems?: ExpenseItemChangesDto;
  settlements?: SettlementChangesDto;
  billMeta?: BillMetaChangesDto;
}