// frontend/src/api/models/memberChangesDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { MemberAddDto } from "./memberAddDto";
import type { MemberUpdateDto } from "./memberUpdateDto";

export interface MemberChangesDto {
  /** @nullable */
  add?: MemberAddDto[] | null;
  /** @nullable */
  update?: MemberUpdateDto[] | null;
  /** @nullable */
  delete?: string[] | null;
}