// frontend/src/api/models/expenseChangesDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { ExpenseAddDto } from "./expenseAddDto";
import type { ExpenseUpdateDto } from "./expenseUpdateDto";

export interface ExpenseChangesDto {
  /** @nullable */
  add?: ExpenseAddDto[] | null;
  /** @nullable */
  update?: ExpenseUpdateDto[] | null;
  /** @nullable */
  delete?: string[] | null;
}