// frontend/src/api/models/expenseItemChangesDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { ExpenseItemAddDto } from "./expenseItemAddDto";
import type { ExpenseItemUpdateDto } from "./expenseItemUpdateDto";

export interface ExpenseItemChangesDto {
  /** @nullable */
  add?: ExpenseItemAddDto[] | null;
  /** @nullable */
  update?: ExpenseItemUpdateDto[] | null;
  /** @nullable */
  delete?: string[] | null;
}