// frontend/src/api/models/settlementChangesDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { DeltaSettlementDto } from "./deltaSettlementDto";

export interface SettlementChangesDto {
  /** @nullable */
  mark?: DeltaSettlementDto[] | null;
  /** @nullable */
  unmark?: DeltaSettlementDto[] | null;
}