// frontend/src/api/models/memberAddDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */

export interface MemberAddDto {
  /** @nullable */
  localId?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  displayOrder?: number | null;
  /** @nullable */
  linkedUserId?: string | null;
  /** @nullable */
  claimedAt?: string | null;
}